refactor structure app
