--- lib/features/blogs/data/models/model.dart
@@ -1,6 +1,0 @@
-import 'package:candidate_app/features/blogs/domain/entities/blog.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'model.freezed.dart';
-part 'model.g.dart';
-part 'blog_model.dart';
