--- lib/features/blogs/domain/repository/blog_repository.dart
@@ -1,9 +1,0 @@
-import 'package:candidate_app/core/error/failures.dart';
-import 'package:dartz/dartz.dart';
-import 'package:kt_dart/collection.dart';
-
-import '../entities/blog.dart';
-
-abstract class BlogRepository {
-  Future<Either<Failure, KtList<Blog>>> getBlogs({String? query});
-}
