--- lib/features/candidates/data/models/model.dart
@@ -1,6 +1,0 @@
-import 'package:candidate_app/features/candidates/domain/entities/candidate.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'candidate_model.dart';
-part 'model.freezed.dart';
-part 'model.g.dart';
