+++ lib/features/home/data/models/model.dart
@@ -1,0 +1,8 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+import '../../domain/entities/entity.dart';
+
+part 'candidate/candidate_model.dart';
+part 'blog/blog_model.dart';
+part 'model.freezed.dart';
+part 'model.g.dart';
