+++ lib/features/home/domain/entities/entity.dart
@@ -1,0 +1,6 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:intl/intl.dart';
+
+part 'blog.dart';
+part 'candidate.dart';
+part 'entity.freezed.dart';
