+++ lib/features/home/domain/usecases/get_candidates.dart
@@ -1,0 +1,30 @@
+import 'package:candidate_app/core/error/failures.dart';
+import 'package:candidate_app/core/usecases/usecase.dart';
+import 'package:dartz/dartz.dart';
+import 'package:equatable/equatable.dart';
+import 'package:injectable/injectable.dart';
+import 'package:kt_dart/collection.dart';
+
+import '../entities/entity.dart';
+import '../repositories/candidate_repository.dart';
+
+@injectable
+class GetCandidates implements Usecase<KtList<Candidate>, Params> {
+  final CandidateRepository _repository;
+
+  GetCandidates(this._repository);
+
+  @override
+  Future<Either<Failure, KtList<Candidate>>> call(params) async {
+    return await _repository.getCandidates(query: params.query);
+  }
+}
+
+class Params extends Equatable {
+  final String? query;
+
+  const Params(this.query);
+
+  @override
+  List<Object?> get props => [query];
+}
